Moved scripts into scripts folder.
Rewired the scene.
