+++ Not-a-gun/Assets/Scripts/ComboSwitch.cs
