+++ Not-a-gun/Assets/Scripts/GateOpen.cs
