+++ Not-a-gun/Assets/Scripts/TriggerCondition.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+interface ITriggerCondition {
+    bool GetTriggered();
+}
