+++ Not-a-gun/Assets/Scripts/TriggerCondition.cs.meta
