+++ Not-a-gun/Assets/Scripts/TriggerProximity.cs.meta
