--- Not-a-gun/Assets/TriggerCondition.cs
@@ -1,7 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-interface ITriggerCondition {
-    bool GetTriggered();
-}
