--- Not-a-gun/Assets/correctheight.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class correctheight : MonoBehaviour {
-
-	// Use this for initialization
-	void Start () {
-        this.transform.position = new Vector3(0, -0.5f, 0);
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
-}
